refactor : Cards라는 player가 가지고 있는 카드뭉치 생성
